controller/requirements.txt
pyaml
dearpygui
obsws_python